refactor into email_utils
