+++ main.py
@@ -1,60 +1,21 @@
 import os.path
 import base64
-import spacy
-from spacy_cleaner import processing, Cleaner
 
 from bs4 import BeautifulSoup
 import re
 
-from google.auth.exceptions import RefreshError
-from google.auth.transport.requests import Request
-from google.oauth2.credentials import Credentials
-from google_auth_oauthlib.flow import InstalledAppFlow
+
 from googleapiclient.discovery import build
 from googleapiclient.errors import HttpError
 
 from db_utils import write_emails
-from email_utils import clean_email, get_word_frequency
+from email_utils import clean_email, get_word_frequency, get_gmail_credentials
 import datetime
 
-# If modifying these scopes, delete the file token.json.
-SCOPES = ["https://www.googleapis.com/auth/gmail.readonly"]
 
 JOBS_LABEL_ID = "Label_7646018251861665561"
 # ideally would be able to programmatically fetch job application-related emails but in interest of time,
 # I manually filtered and placed in this label 'jobs' with this id starting with Label_
-
-
-def get_gmail_credentials():
-    creds = None
-    # The file token.json stores the user's access and refresh tokens, and is
-    # created automatically when the authorization flow completes for the first
-    # time.
-    if os.path.exists("token.json"):
-        creds = Credentials.from_authorized_user_file("token.json", SCOPES)
-    # If there are no (valid) credentials available, let the user log in.
-
-    if not creds or not creds.valid:
-        if creds and creds.expired and creds.refresh_token:
-            try:
-                creds.refresh(Request())
-            except RefreshError:
-                os.remove("token.json")
-                creds.refresh(Request())
-        else:
-            try:
-                flow = InstalledAppFlow.from_client_secrets_file(
-                    "credentials.json",
-                    scopes=SCOPES,
-                )
-                creds = flow.run_local_server(port=8001)
-            except RefreshError:
-                os.remove("token.json")
-                creds.refresh(Request())
-        # Save the credentials for the next run
-        with open("token.json", "w") as token:
-            token.write(creds.to_json())
-    return creds
 
 
 def main():
